refactor: 고차함수 폴더구조 변경, util/form 의 내용을 useAutoFocusForm으로 이동
