+++ src/helper/withRouteState.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { useLocation } from 'react-router-dom';
-import ErrorPage from '../../../pages/404';
-
+import ErrorPage from '../pages/404';
 function withRouteState(Component) {
   const UseRouteStateComponent = () => {
     const { state } = useLocation();
