+++ src/pages/cardlist/success/index.jsx
@@ -3,7 +3,7 @@ import PageTemplate from '../../../components/commons/PageTemplate';
 import Card from '../../../components/card/Card';
 import AliasInputForm from '../../../components/form/InputForm/AliasInputForm';
 import PropTypes from 'prop-types';
-import withRouteState from '../../../components/commons/withRouteState';
+import withRouteState from '../../../helper/withRouteState';
 
 function CardSuccess({ cardListDispatch, getCard, routeState: { cardId } }) {
   const card = getCard(cardId);
