--- src/utils/util/form.js
@@ -1,11 +1,0 @@
-export const findNotCompletedInput = (inputMap, currentInputKey) => {
-  const nextInputKey = Object.keys(inputMap).find(
-    key => key !== currentInputKey && inputMap[key].isComplete === false,
-  );
-
-  return (
-    inputMap[nextInputKey] ?? {
-      element: null,
-    }
-  );
-};
